extract to class
